Quick fix for sleep timers
The problem was caused by the state transition stepping on the `lastRun`
value. Was able to implement a quick fix by reordering these operations.
This is a hacky solution and a more solid one would be to remove the
`lastRun` value out of the scraper and into the scraper's state machine.